# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentStamp API",
    "version": "2.3.0",
    "description": "AgentStamp provides cryptographic identity certificates (stamps), a searchable public registry, reputation scores, cross-protocol passports, verification badges, and a Wishing Well marketplace for AI agents. Free tier available. Paid endpoints use x402 micropayments (USDC on Base and Solana).",
    "contact": {
      "name": "AgentStamp",
      "url": "https://agentstamp.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://agentstamp.org",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/webhooks/register": {
      "post": {
        "operationId": "registerWebhook",
        "summary": "Register a webhook",
        "description": "Register a webhook to receive notifications for events. Maximum 3 webhooks per wallet.",
        "tags": ["Webhooks"],
        "parameters": [
          { "$ref": "#/components/parameters/XWalletAddress" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRegisterBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Duplicate webhook URL for this wallet",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Maximum 3 webhooks per wallet",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks": {
      "get": {
        "operationId": "listWebhooks",
        "summary": "List webhooks",
        "description": "List all webhooks registered for a wallet. Requires x-wallet-address header.",
        "tags": ["Webhooks"],
        "parameters": [
          {
            "name": "x-wallet-address",
            "in": "header",
            "required": true,
            "description": "Wallet address to list webhooks for",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListResponse"
                }
              }
            }
          },
          "400": {
            "description": "x-wallet-address header required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{id}": {
      "delete": {
        "operationId": "deleteWebhook",
        "summary": "Delete a webhook",
        "description": "Delete a webhook by ID. Only the webhook owner can delete it.",
        "tags": ["Webhooks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The webhook ID to delete",
            "schema": { "type": "string" }
          },
          {
            "name": "x-wallet-address",
            "in": "header",
            "required": true,
            "description": "Wallet address of the webhook owner",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "x-wallet-address header required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Only the webhook owner can delete it",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Webhook not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "XWalletAddress": {
        "name": "x-wallet-address",
        "in": "header",
        "description": "Wallet address (EVM or Solana). Can also be provided as wallet_address in the request body.",
        "schema": { "type": "string" }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" }
        }
      },
...
      "SuccessMessageResponse": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "message": { "type": "string" }
        }
      },
...
      "WebhookRegisterBody": {
        "type": "object",
        "required": ["wallet_address", "url", "events"],
        "properties": {
          "wallet_address": { "type": "string" },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "HTTPS URL to receive webhook POST requests"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["stamp_minted", "stamp_expiring", "endorsement_received", "wish_granted", "wish_matched", "reputation_changed", "agent_registered"]
            },
            "minItems": 1,
            "description": "Events to subscribe to"
          }
        }
      },
      "WebhookRegisterResponse": {
        "type": "object",
        "required": ["success", "webhook", "note"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "webhook": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "url": { "type": "string" },
              "events": {
                "type": "array",
                "items": { "type": "string" }
              },
              "secret": {
                "type": "string",
                "description": "HMAC-SHA256 secret for verifying webhook payloads (X-AgentStamp-Signature header)"
              }
            }
          },
          "note": { "type": "string" }
        }
      },
      "WebhookListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "events": {
            "type": "array",
            "items": { "type": "string" }
          },
          "active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "WebhookListResponse": {
        "type": "object",
        "required": ["success", "webhooks"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "webhooks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WebhookListItem" }
          }
        }
      },
...
    }
  }
}